lambda/requirements.txt
pymongo==3.12.0
opentelemetry-instrumentation-urllib3